chore: deprecates v1
